+++ src/pages/knowledgegraph/DetailsPopup/EntityTypeDetails/index.ts
