+++ src/pages/knowledgegraph/DetailsPopup/ImageDetails/ImageDetails.tsx
