+++ src/pages/knowledgegraph/DetailsPopup/ImageDetails/index.ts
@@ -1,0 +1,1 @@
+export * from './ImageDetails';
