--- src/pages/knowledgegraph/SelectionDetails/index.ts
@@ -1,1 +1,0 @@
-export * from './SelectionDetails';
